feat: add Router & Button Link
Add react dom router and fix the warning of the main page
