+++ frontend/src/index.js
@@ -2,11 +2,9 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
-import Header from "./container/Header";
 
 ReactDOM.render(
   <div>
-    <Header />
     <App />
   </div>,
   document.getElementById("root")
